requirements.txt
python-dotenv==1.1.1
watercrawl-py==0.9.2
